fix: add existance check

+++ voicevox_engine/sv_model.py
@@ -80,7 +80,7 @@ def register_sv_model(
             speaker_info_dir = stored_dir / "speaker_info" / speaker_uuid
             
             # 既にモデルが存在していた場合はrenameしておく
-            if already_exists:
+            if already_exists and os.path.exists(speaker_info_dir):
                 os.rename(speaker_info_dir, f"{speaker_info_dir}.old")
 
             os.makedirs(speaker_info_dir / "icons")
@@ -126,7 +126,9 @@ def register_sv_model(
             shutil.rmtree(f"{model_uuid_dir}.old")
             for speaker_uuid, speaker_info in sv_model.speaker_infos.items():
                 speaker_info_dir = stored_dir / "speaker_info" / speaker_uuid
-                shutil.rmtree(f"{speaker_info_dir}.old")
+                # 新しく追加されるspeaker_info_dirに.oldは存在しないはずなので、exists checkをする
+                if os.path.exists(f"{speaker_info_dir}.old"):
+                    shutil.rmtree(f"{speaker_info_dir}.old")
 
     except Exception as e:
         # 削除時にエラーが発生しても無視する
@@ -140,5 +142,6 @@ def register_sv_model(
         os.rename(f"{model_uuid_dir}.old", model_uuid_dir)
         for speaker_uuid, speaker_info in sv_model.speaker_infos.items():
             speaker_info_dir = stored_dir / "speaker_info" / speaker_uuid
-            os.rename(f"{speaker_info_dir}.old", speaker_info_dir)
+            if os.path.exists(f"{speaker_info_dir}.old"):
+                os.rename(f"{speaker_info_dir}.old", speaker_info_dir)
         raise e
